Handle imgur links without extensions
Previously only links to i.imgur.com would be inlined—links to pages
containing images would be ignored, even though they are often the links
given by "share" buttons. @ViMaSter registered an API client and
authored changes to inline album links (imgur.com/a/{ID}) by inlining
the first image in the album—this commit expands on this to inline image
and gallery links (imgur.com/{ID} and imgur.com/gallery/{ID}) too,
and to support inlining videos and images in formats other than PNG.

Because the imgur API is used to acquire the direct link a rate limit of
12,500 requests per day exists, which if hit 5 times in a month will be
reduced to 0 for the remainder of the month. Hopefully this doesn't turn
out to be problem.

Closes #8.

Co-authored-by: Vincent Mahnke <accounts@vimaster.de>