Actually fix imgur proxy race conditions
This is why I shouldn't rush to push an update in time for Desert Bus...

I originally awaited the imgur blocked test after starting to observe
for the chat window to try and prevent us from missing messages while
waiting for a response from imgur... but I can just check for existing
messages in the chat window when we find it!

This does mean that no messages are inlined until imgur responds, but
fixing that means rewriting this script to be much more asynchronous,
something I definitely do not want to try and rush out.

Actually fixes #19.